refactor(ocr): move runtime/session init into engine/runtime
